+++ packages/std-client/src/setup/setupMUDNetwork.ts
@@ -5,13 +5,16 @@ import {
   createTxQueue,
   createSyncWorker,
   createEncoder,
-  NetworkComponentUpdate,
   createSystemExecutor,
   NetworkConfig,
   SyncWorkerConfig,
   isNetworkComponentUpdateEvent,
+  NetworkEvent,
+  ack,
+  Ack,
+  InputType,
 } from "@latticexyz/network";
-import { BehaviorSubject, bufferTime, filter, Observable, Subject } from "rxjs";
+import { BehaviorSubject, concatMap, filter, from, map, Observable, Subject, timer } from "rxjs";
 import {
   Component,
   Components,
@@ -25,7 +28,7 @@ import {
   World,
 } from "@latticexyz/recs";
 import { computed, IComputedValue } from "mobx";
-import { keccak256, stretch, toEthAddress } from "@latticexyz/utils";
+import { keccak256, toEthAddress } from "@latticexyz/utils";
 import ComponentAbi from "@latticexyz/solecs/abi/Component.json";
 import { Contract, ContractInterface, Signer } from "ethers";
 import { Component as SolecsComponent } from "@latticexyz/solecs";
@@ -38,7 +41,6 @@ export type SetupContractConfig = NetworkConfig &
   Omit<SyncWorkerConfig, "worldContract" | "mappings"> & {
     worldAddress: string;
     devMode?: boolean;
-    limitEventsPerSecond?: number;
   };
 
 export type ContractComponent = Component<Schema, { contractId: string }>;
@@ -106,29 +108,34 @@ export async function setupMUDNetwork<C extends ContractComponents, SystemTypes 
   });
   world.registerDisposer(disposeTxQueue);
 
-  const systems = createSystemExecutor<SystemTypes>(world, network, SystemsRegistry, SystemAbis, gasPriceInput$, {
-    devMode: networkConfig.devMode,
-  });
+  const { systems, registerSystem } = createSystemExecutor<SystemTypes>(
+    world,
+    network,
+    SystemsRegistry,
+    SystemAbis,
+    gasPriceInput$,
+    {
+      devMode: networkConfig.devMode,
+    }
+  );
 
   // Create sync worker
-  const { ecsEvent$, config$, dispose } = createSyncWorker<C>();
+  const ack$ = new Subject<Ack>();
+  const { ecsEvents$, input$, dispose } = createSyncWorker<C>(ack$);
   world.registerDisposer(dispose);
   function startSync() {
-    config$.next({
-      ...networkConfig,
-      worldContract: contractsConfig.World,
-      initialBlockNumber: networkConfig.initialBlockNumber ?? 0,
-      disableCache: networkConfig.devMode, // Disable cache on local networks (hardhat / anvil)
+    input$.next({
+      type: InputType.Config,
+      data: {
+        ...networkConfig,
+        worldContract: contractsConfig.World,
+        initialBlockNumber: networkConfig.initialBlockNumber ?? 0,
+        disableCache: networkConfig.devMode, // Disable cache on local networks (hardhat / anvil)
+      },
     });
   }
 
-  const { txReduced$ } = applyNetworkUpdates(
-    world,
-    components,
-    ecsEvent$.pipe(filter(isNetworkComponentUpdateEvent)),
-    mappings,
-    networkConfig.limitEventsPerSecond
-  );
+  const { txReduced$ } = applyNetworkUpdates(world, components, ecsEvents$, mappings, ack$);
 
   const encoders = networkConfig.encoders
     ? createEncoders(world, ComponentsRegistry, signerOrProvider)
@@ -142,9 +149,10 @@ export async function setupMUDNetwork<C extends ContractComponents, SystemTypes 
     startSync,
     systems,
     gasPriceInput$,
-    ecsEvent$,
+    ecsEvent$: ecsEvents$.pipe(concatMap((updates) => from(updates))),
     mappings,
     registerComponent,
+    registerSystem,
   };
 }
 
@@ -184,39 +192,49 @@ async function createEncoders(
 function applyNetworkUpdates<C extends Components>(
   world: World,
   components: C,
-  ecsEvent$: Observable<NetworkComponentUpdate<C>>,
+  ecsEvents$: Observable<NetworkEvent<C>[]>,
   mappings: Mappings<C>,
-  limitEventsPerSecond = 62500
+  ack$: Subject<Ack>
 ) {
   const txReduced$ = new Subject<string>();
-  const limitEventsPer16ms = Math.floor(limitEventsPerSecond / 62.5);
 
-  const ecsEventSub = ecsEvent$
+  // Send "ack" to tell the sync worker we're ready to receive events while not processing
+  let processing = false;
+  const ackSub = timer(0, 100)
     .pipe(
-      // We throttle the client side event processing to 1000 events every 16ms by default, so 62.500 events per second.
-      // This means if the chain were to emit more than 62.500 events per second, the client would not keep up.
-      // The only time we get close to this number is when initializing from a snapshot/cache.
-      bufferTime(16, null, limitEventsPer16ms),
-      filter((updates) => updates.length > 0),
-      stretch(16)
+      filter(() => !processing),
+      map(() => ack)
     )
-    .subscribe((updates) => {
-      for (const update of updates) {
-        const entityIndex = world.entityToIndex.get(update.entity) ?? world.registerEntity({ id: update.entity });
-        const componentKey = mappings[update.component];
-        if (!componentKey) return console.warn("Unknown component:", update);
+    .subscribe(ack$);
 
-        if (update.value === undefined) {
-          // undefined value means component removed
-          removeComponent(components[componentKey] as Component<Schema>, entityIndex);
-        } else {
-          setComponent(components[componentKey] as Component<Schema>, entityIndex, update.value);
-        }
+  const delayQueueSub = ecsEvents$.subscribe((updates) => {
+    processing = true;
+    for (const update of updates) {
+      if (!isNetworkComponentUpdateEvent<C>(update)) continue;
+      if (update.lastEventInTx) txReduced$.next(update.txHash);
 
-        if (update.lastEventInTx) txReduced$.next(update.txHash);
+      const entityIndex = world.entityToIndex.get(update.entity) ?? world.registerEntity({ id: update.entity });
+      const componentKey = mappings[update.component];
+      if (!componentKey) {
+        console.warn("Unknown component:", update);
+        continue;
       }
-    });
 
-  world.registerDisposer(() => ecsEventSub?.unsubscribe());
+      if (update.value === undefined) {
+        // undefined value means component removed
+        removeComponent(components[componentKey] as Component<Schema>, entityIndex);
+      } else {
+        setComponent(components[componentKey] as Component<Schema>, entityIndex, update.value);
+      }
+    }
+    // Send "ack" after every processed batch of events to process faster than ever 100ms
+    ack$.next(ack);
+    processing = false;
+  });
+
+  world.registerDisposer(() => {
+    delayQueueSub?.unsubscribe();
+    ackSub?.unsubscribe();
+  });
   return { txReduced$: txReduced$.asObservable() };
 }